// apps/web/components/record-media-selected-content.types.ts
import type { RecordMediaToolsProps } from "./record-media-tools.types";

export type RecordMediaSelectedContentProps = Pick<
  RecordMediaToolsProps,
  | "allTrackedFilesPresentLabel"
  | "authToken"
  | "bulkRetryingDeadLetter"
  | "canWriteWorkspace"
  | "deletingMediaId"
  | "downloadingMediaId"
  | "formatFileCountLabel"
  | "formatHistoryTimestampLabel"
  | "largestFilePrefixLabel"
  | "locale"
  | "localLabel"
  | "mediaAssets"
  | "mediaDeadLetterOverview"
  | "mediaIssueCopy"
  | "mediaProcessingOverview"
  | "mediaStorageSummary"
  | "missingFilesLabel"
  | "needsAttentionLabel"
  | "noMediaLabel"
  | "onBulkRetryAllDeadLetter"
  | "onBulkRetrySelectedDeadLetter"
  | "onClearDeadLetterSelection"
  | "onDeleteMediaAsset"
  | "onDownloadMedia"
  | "onRefreshMedia"
  | "onRetryMediaProcessing"
  | "onSelectAllDeadLetter"
  | "onToggleDeadLetterSelection"
  | "processingCompletedLabel"
  | "queuedLabel"
  | "queueStateLabel"
  | "refreshingMediaId"
  | "remoteLabel"
  | "retryingMediaId"
  | "selectedDeadLetterIds"
  | "selectedRecordMediaSizeLabel"
  | "storageHealthLabel"
  | "storageMixLabel"
  | "thisRecordMediaLabel"
  | "workspaceId"
  | "workspaceStorageLabel"
>;